Alert before delete - mockup

+++ app/instrument/challenge/index/controller.js
@@ -2,8 +2,29 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
   actions: {
+    confirmWithUser() {
+      var response = confirm('Are you sure you want to delete this challenge?');
+      return response;
+    },
+
     deleteChallenge(challenge) {
-      challenge.destroyRecord();
+      const response = confirmWithUser();
+      if(response) {
+        challenge.destroyRecord()
+      }
+      else {
+        this.transitionToRoute('instrument.challenge');
+      }
     }
+    // deleteChallenge(challenge) {
+    //
+    //   if(alert = true) {
+    //     challenge.destroyRecord();
+    //   }
+    //
+    //   else {
+    //     this.transitionToRoute('instrument.challenge')
+    //   }
+    // }
   }
 });
